fix [박현우] : 바텀시트모달 로직 변경 및 사용법 작성
